extraction-backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pytesseract
pillow
pypdf
pypdfium2
